backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.7.0
bcrypt==4.0.1
supabase==2.4.0